# portless-daemon/api/openapi.yaml
openapi: 3.1.0
info:
  title: Portless local control API
  version: 10.7.0
  description: >-
    Loopback-only API shared by the Portless CLI and embedded UI. Projects are
    reusable application topologies. Environments are independently configured
    and runnable instances. Public routes use readable project, environment,
    service, and artifact names; private ownership keys are never exposed.
    Authenticated native clients may send Portless-Client-Kind as `cli` or
    `mcp`; the daemon records that fixed actor category on durable operations.
    Browser sessions cannot claim a native client kind.
servers:
  - url: http://portless.localhost/api/v1
security:
  - bearer: []
  - browserSession: []
paths:
...
  /runtime/start:
    post:
      operationId: startRuntime
      responses:
        '200': { description: Selected container runtime is ready }
        '503': { $ref: '#/components/responses/Unavailable' }
...
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
      description: Mode-0600 installation secret used by the local CLI.
    browserSession:
      type: apiKey
      in: cookie
      name: portless_session
      description: SameSite=Strict browser session; mutations also require X-Portless-CSRF.
...
  responses:
...
    Unavailable:
      description: Required local runtime unavailable
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
            subject: { type: object, additionalProperties: true }
            details: { type: object, additionalProperties: true }
            remediation: { type: array, items: { type: object, additionalProperties: true } }